- fix usage of InterlockedExchangePointer - some ULONG / SIZE_T fixes - pushlocks have a union with a PVOID and an ULONG, use them instead of typecasting - use InterlockedAndPointer for a pushlock
svn path=/branches/ros-amd64-bringup/; revision=34986

+++ reactos/ntoskrnl/se/lsa.c
@@ -49,7 +49,7 @@ NTSTATUS
 NTAPI
 LsaFreeReturnBuffer(PVOID Buffer)
 {
-    ULONG Size = 0;
+    SIZE_T Size = 0;
     return ZwFreeVirtualMemory(NtCurrentProcess(),
                                &Buffer,
                                &Size,
